College_chatbot/requirements.txt
flask==3.0.0
spacy
python-dotenv==1.0.0
